requirements.txt
numpy
aiohttp
openai
uvicorn
fastapi
pillow
pydantic
gradio
requests
python-dotenv
google-cloud-storage
google-auth
chardet
groq
python-json-logger
apscheduler
matplotlib
perlin-noise
loguru